% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioview-package.R
\docType{package}
\name{biblioview-package}
\alias{biblioview}
\alias{biblioview-package}
\title{biblioview: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tpetzoldt/biblioview}
  \item Report bugs at \url{https://github.com/tpetzoldt/biblioview/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
