% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_base_req.R
\name{build_base_req}
\alias{build_base_req}
\title{Build Base Request Template with Polite Headers and Retries}
\usage{
build_base_req(email_contact)
}
\arguments{
\item{email_contact}{Character. Email used for polite API headers.}
}
\value{
An httr2 request object.
}
\description{
Build Base Request Template with Polite Headers and Retries
}
\keyword{internal}
