% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_abstract.R
\name{fetch_abstract_crossref}
\alias{fetch_abstract_crossref}
\title{Fetch Abstract from Crossref}
\usage{
fetch_abstract_crossref(doi, req_template)
}
\arguments{
\item{doi}{Clean DOI string.}

\item{req_template}{httr2 base request.}
}
\value{
Character string containing abstract or NULL if not found.
}
\description{
Fetch Abstract from Crossref
}
\keyword{internal}
