% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_abstract.R
\name{fetch_abstract_epmc}
\alias{fetch_abstract_epmc}
\title{Fetch Abstract from Europe PMC}
\usage{
fetch_abstract_epmc(doi, req_template)
}
\arguments{
\item{doi}{Clean DOI string.}

\item{req_template}{httr2 base request.}
}
\value{
Character string containing abstract or NULL if not found.
}
\description{
Fetch Abstract from Europe PMC
}
\keyword{internal}
