% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ris_from_crossref.R
\name{fetch_ris_from_crossref}
\alias{fetch_ris_from_crossref}
\title{Fetch Bibliography Metadata in RIS Format from Crossref}
\usage{
fetch_ris_from_crossref(dois, email = NULL)
}
\arguments{
\item{dois}{Character vector of DOIs.}

\item{email}{Optional email address for Crossref's polite pool (recommended).}
}
\value{
A character vector of RIS formatted entries.
}
\description{
Takes a vector of DOIs and retrieves their bibliographic data formatted as RIS
using Crossref Content Negotiation.
}
\examples{
\dontrun{
fetch_ris_from_crossref(c("10.1016/j.ecolmodel.2020.109282", "10.1111/gcb.15000"))
}
}
