% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioview-package.R
\docType{package}
\name{biblioview-package}
\alias{biblioview}
\alias{biblioview-package}
\title{biblioview: A Shiny Web Portal and Abstract Enricher for Zotero Libraries}
\description{
Provides automated pipelines to fetch Zotero group libraries, generate APA citations, and harvest missing abstracts from external metadata APIs (Crossref, OpenAlex, and Europe PMC).
}
\author{
\strong{Maintainer}: Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de}

Authors:
\itemize{
  \item Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de}
}

}
\keyword{internal}
