% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioview-package.R
\docType{package}
\name{biblioview-package}
\alias{biblioview}
\alias{biblioview-package}
\title{biblioview: A Shiny Web Portal and Abstract Enricher for Zotero Libraries}
\description{
Provides automated pipelines to fetch Zotero group libraries, generate APA citations, and harvest citation counts and missing abstracts from external metadata APIs (Crossref, OpenAlex, and Europe PMC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tpetzoldt.github.io/biblioview/}
}

}
\author{
\strong{Maintainer}: Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de} (\href{https://orcid.org/0000-0002-4951-6468}{ORCID})

Authors:
\itemize{
  \item Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de} (\href{https://orcid.org/0000-0002-4951-6468}{ORCID})
}

}
\keyword{internal}
