% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_missing_abstracts.R
\name{enrich_missing_abstracts}
\alias{enrich_missing_abstracts}
\title{Enrich Missing Abstracts from External Metadata APIs}
\usage{
enrich_missing_abstracts(df, email_contact = Sys.getenv("POLITE_EMAIL"))
}
\arguments{
\item{df}{A data frame generated by \code{fetch_all_zotero_data}.}

\item{email_contact}{Character. A valid email address used for the polite API pool.
Defaults to the environment variable \code{POLITE_EMAIL}.}
}
\description{
Enrich Missing Abstracts from External Metadata APIs
}
