% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_abstract.R
\name{fetch_abstract_semanticscholar}
\alias{fetch_abstract_semanticscholar}
\title{Fetch Abstract from Semantic Scholar}
\usage{
fetch_abstract_semanticscholar(doi, base_req)
}
\arguments{
\item{doi}{Clean DOI string.}

\item{base_req}{httr2 base request.}
}
\value{
Character string containing abstract or NULL if not found.
}
\description{
Fetch Abstract from Semantic Scholar
}
